+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Work In Progress
 - Addition and subtraction of Tranforms supported to help with easing function calculations
+- Added an example with loading progress bar
 
 ## v0.4.0-alpha0.3
 - Update `golem` to `v0.1.1` to fix non-power-of-2 textures