+++ CHANGES.md
@@ -3,6 +3,8 @@
 ## Work In Progress
 - Addition and subtraction of Tranforms supported to help with easing function calculations
 - Added an example with loading progress bar
+- `lifecycle::run` can now accept any kind of Error.
+- Add `into_raw_context` on Graphics, to allow lower-level graphics programming
 
 ## v0.4.0-alpha0.3
 - Update `golem` to `v0.1.1` to fix non-power-of-2 textures