+++ CHANGES.md
@@ -9,6 +9,7 @@
 ## v0.4.0-alpha0.3
 - Update `golem` to `v0.1.1` to fix non-power-of-2 textures
 - `impl std::iter::Sum for geom::Vector`
+- Implement `std::ops::MulAssign`, `std::ops::AddAssign`, and `std::ops::SubAssign` for `Transform`s.
 
 ## v0.4.0-alpha0.2
 - Fix the "easy-log" feature