+++ CHANGES.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## WIP
+- Fix compile issues with font-related features
+
 ## v0.4.0-alpha0.3
 - Update `golem` to `v0.1.1` to fix non-power-of-2 textures
 - `impl std::iter::Sum for geom::Vector`